pyproject.toml
[project]
name = "supernote-apple-note-unifier"
version = "0.1.0"
description = "Mirror Apple Notes to Supernote as .note files"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "pillow>=10.0",
    "reportlab>=4.0",
    "supernotelib>=0.4.0",
    "pypdf>=4.0",
    "pymupdf>=1.24.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
unifier = "unifier.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/unifier"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
